services/fraud-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
scipy==1.11.4

# Anomaly detection
pyod==1.1.0

# Stream processing
confluent-kafka==2.3.0

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0